Moves profile route to seperate file
